requirements.txt
fastapi[standard]>=0.115.11
uvicorn>=0.34.0
httpx>=0.25.0
pydantic>=2.0.0
litellm>=1.40.14
python-dotenv>=1.0.0
openai
mcp
json-repair